refactor(audit-labellisation): édite les notes de l'auditeur en inline plutôt qu'en side-panel
Aligne la cellule des notes d'audit sur la cellule d'état d'avancement :
remplace l'ouverture d'un side-panel (RichTextEditor + useSidePanel) par une
édition directe via TableCellRichTextEditor. Le helper de preview tronqué,
devenu sans usage, est supprimé avec son spec.
